requirements.txt
# Python AI Agentic Framework for OSINT Collection
# LangGraph for workflow orchestration
langgraph==0.2.45
langgraph-checkpoint==2.0.6

# Anthropic API for Claude integration
anthropic==0.45.0

# Environment variable management
python-dotenv==1.0.1

# Additional utilities
